requirements.txt
--index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
dotenv==0.9.9
filelock==3.18.0
fsspec==2025.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.4.2
notion-client==2.3.0
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
pillow==11.0.0
python-dotenv==1.1.0
pytube==15.0.0
regex==2024.11.6
requests==2.32.3
setuptools==80.8.0
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yt-dlp==2025.5.22
